Internal functions exports added to fix TFCE output discrepancies.
When the package is simply loaded with library(VertexWiseR), non-exported internal functions are not properly accessed by the vertTFCE, which caused spurious TFCE results. I'm not sure why they have to be exported as they normally don't need to if inside the package.

getFromNamespace avoids using ::: which CRAN rejects, and source(otherfunc.r) which is discouraged (https://r-pkgs.org/code.html). Source also uses URLs which requires being online, or full paths which can't be predefined for the user.

See:
https://stat.ethz.ch/pipermail/r-help/2007-May/132366.html
https://forum.posit.co/t/how-to-include-internal-functions/38654
https://stackoverflow.com/questions/22809758/r-source-command-in-package
https://stackoverflow.com/questions/34589497/r-calling-internal-function-from-an-exported-function-inside-package