fixed 'NoneType' object is not subscriptable
'NoneType' object in python indicates an invalid object. It appears that previously hip_points_cells was not correctly loaded. Not sure why, could be some environment-related issue.

For some reason if get('hip_points_cells') is fed directly to the hipdat argument in surfplot_canonical_foldunfold() it works. Maybe it has something to do with bypassing the environment